feat : move logic to orderList component
