+++ node/package.json
@@ -1,6 +1,6 @@
 {
   "dependencies": {
-    "@vtex/api": "6.45.10",
+    "@vtex/api": "6.45.11-beta",
     "@vtex/clients": "^2.13.0",
     "axios": "^0.20.0",
     "co-body": "^6.0.0",
@@ -13,7 +13,7 @@
     "@types/jest": "^24.0.18",
     "@types/node": "^12.0.0",
     "@types/ramda": "types/npm-ramda#dist",
-    "@vtex/api": "6.45.10",
+    "@vtex/api": "6.45.11-beta",
     "@vtex/test-tools": "^1.0.0",
     "@vtex/tsconfig": "^0.6.0",
     "tslint": "^6.1.3",
